+++ packages/queue/src/queue/jobs/index.ts
@@ -1,0 +1,7 @@
+export {
+    createJob,
+    InvalidJobOptionError,
+    isJob,
+    type CreateJobOptions,
+    type Job,
+} from './job'